requirements.txt
PySide6>=6.10.2
requests>=2.32.5
cloudscraper-enhanced>=3.0.0
Js2Py-3.13>=0.74.1